Initial commit of the composer setup
Comitting the composer (json and lock) files so that robots have the instructions to build the site at my whim. Changing the  .gitignore file to play nicely with the site's robot overlords. Using the .gitattributes file created by composer because that seems more appropriate to the job than the default file that GitHub provided.