Refactoring theme
Regenerated (via Drush) to create a new base theme from scratch that uses Bartik as a base. Added support for the color module by copying those files from Bartik. This was done as a step to get the site to a more presentable stage, quicker. More customizations will be made down the road.